backend/requirements.prod.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
greenlet==3.0.1  # Required by SQLAlchemy async

# HTTP client
httpx==0.27.2
aiohttp==3.9.1

# Scheduling
apscheduler==3.10.4

# Async job queue (Cloudflare Workers + Upstash Redis)
upstash-redis==0.15.0

# Authentication
supabase==2.9.1
pydantic-settings==2.1.0

# String matching and fuzzy search
rapidfuzz==3.5.2

# Background jobs and caching (required by production services)
celery==5.3.4
redis==5.0.1

# Minimal dependencies for production only
# Removed: pytest, mypy, ruff, selenium, PyMuPDF (dev/test only)
# Added back: rapidfuzz (citation_service, paper_crawler), celery (jobs), redis (cache_service)
